requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aniso8601==8.0.0          # via flask-restx
attrs==19.3.0             # via automat, jsonschema, service-identity, twisted
automat==20.2.0           # via twisted
cffi==1.14.0              # via cryptography
click==7.1.2              # via flask
constantly==15.1.0        # via twisted
crochet==1.12.0           # via -r requirements.in
cryptography==2.9.2       # via pyopenssl, scrapy, service-identity
cssselect==1.1.0          # via parsel, scrapy
flask-restx==0.2.0        # via -r requirements.in
flask==1.1.2              # via -r requirements.in, flask-restx
hyperlink==19.0.0         # via twisted
idna==2.10                # via hyperlink
importlib-metadata==1.7.0  # via jsonschema
incremental==17.5.0       # via twisted
itemadapter==0.1.0        # via scrapy
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
jsonschema==3.2.0         # via flask-restx
lxml==4.5.1               # via parsel, scrapy
markupsafe==1.1.1         # via jinja2
parsel==1.6.0             # via scrapy
protego==0.1.16           # via scrapy
pyasn1-modules==0.2.8     # via service-identity
pyasn1==0.4.8             # via pyasn1-modules, service-identity
pycparser==2.20           # via cffi
pydispatcher==2.0.5       # via scrapy
pyhamcrest==2.0.2         # via twisted
pyopenssl==19.1.0         # via scrapy
pyrsistent==0.16.0        # via jsonschema
pytz==2020.1              # via flask-restx
queuelib==1.5.0           # via scrapy
scrapy==2.2.0             # via -r requirements.in
service-identity==18.1.0  # via scrapy
six==1.15.0               # via automat, cryptography, flask-restx, jsonschema, parsel, protego, pyopenssl, pyrsistent, w3lib
twisted==20.3.0           # via crochet, scrapy
w3lib==1.22.0             # via parsel, scrapy
werkzeug==1.0.1           # via -r requirements.in, flask, flask-restx
wrapt==1.12.1             # via crochet
zipp==3.1.0               # via importlib-metadata
zope.interface==5.1.0     # via scrapy, twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
